% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mtp.R
\name{as_mtp}
\alias{as_mtp}
\title{Converts a data frame to a data frame of class mtp}
\usage{
as_mtp(df, type = "raw")
}
\arguments{
\item{df}{a data frame to convert. It must be a data frame containing the
following variables: \code{Hor.}, \code{Ver.}, \code{Name}, \code{hours},
and \code{value}.}

\item{type}{is an attribute to be set to show what kind of manipulations have been done (currently not used)}
}
\value{
Returns a df of class mtp
}
\description{
Converts a data frame to a data frame of class mtp
}
\examples{
df_mtp <- xls_read(path = system.file("extdata", "crystalban_salt_01.xls", package = "RAPr"))
mtp <- df_mtp$mtp
class(df_mtp)
str(df_mtp)
df_mtp
summary(df_mtp)

}

