% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{check_genius_token}
\alias{check_genius_token}
\title{Check if the Genius API token is set or prompt for it}
\usage{
check_genius_token()
}
\value{
None. Sets the environment variable \code{GENIUS_API_TOKEN} if the token is provided interactively.
Throws an error if the token is not set and not provided interactively.
}
\description{
This internal function checks if the Genius API token is set in the environment.
If the token is not set, it prompts the user to enter it interactively.
The token is then stored in the environment for the current R session.
}
\keyword{internal}
