% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mtp.R
\name{combine_mtp_list}
\alias{combine_mtp_list}
\title{Combines \code{mtp} objects in a list}
\usage{
combine_mtp_list(list_of_mtp)
}
\arguments{
\item{list_of_mtp}{A list of \code{mtp} objects}
}
\value{
Returns a list of combined \code{mtp} objects
}
\description{
Combines \code{mtp} objects in a list
}
\examples{
# df_mtp is supplied with the package
combined <- combine_mtp_list(df_mtp$mtp)
plot_mtp_list(combine_mtp_list(df_mtp$mtp))

# using dplyr
library(dplyr)
combined <- df_mtp \%>\% dplyr::mutate(cmtp = combine_mtp_list(.$mtp)) # recycling
combined <- df_mtp \%>\% dplyr::group_by(file) \%>\% dplyr::summarize(cmtp = combine_mtp_list(mtp)) #condensing
plot_mtp_list(combined$cmtp)
}

