% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_mtp.R
\name{contrast_mtp_list}
\alias{contrast_mtp_list}
\title{Subtracts the first \code{mtp} object in a list of \code{mtp} objects from the remaining objects}
\usage{
contrast_mtp_list(list_of_mtp)
}
\arguments{
\item{list_of_mtp}{a list of \code{mtp} objects}
}
\value{
Returns a list of contrasted \code{mtp} objects
}
\description{
Subtracts the first \code{mtp} object in a list of \code{mtp} objects from the remaining objects
}
\examples{
contrast_list <- contrast_mtp_list(df_mtp$mtp[1:2])
plot_mtp_list(contrast_list)
plot_mtp_list(smooth_mtp_list(contrast_list), y_name = "fit")

}

