% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{fun_smooth}
\alias{fun_smooth}
\title{Loess smoothing of a well curve}
\usage{
fun_smooth(df, x_name = "hours", y_name = "value", no_points = NULL,
  fit_degree = 1)
}
\arguments{
\item{df}{takes a \code{mtp} object as input}

\item{x_name}{the name of the x variable of type character}

\item{y_name}{the name of the y variable of type character}

\item{no_points}{The number of points in the loess smooth window of type numeric. Defaults to 20\% of the length of the x variable}

\item{fit_degree}{The degree of the loess smooth polynomial of type numeric must be less than or equal to 2}
}
\value{
returns a \code{mtp} object
}
\description{
Loess smoothing of a well curve
}
\examples{
df <- df_mtp$mtp[[1]]
curve <- subset(df, curve_id == 1)
fun_smooth(curve)
fun_smooth(curve, no_points = 40)
}
\seealso{
\code{\link{smooth_mtp}} for mtp objects and \code{\link{smooth_mtp_list}} for lists of mtp objects

Other smoothing.functions: \code{\link{smooth_mtp_list}},
  \code{\link{smooth_mtp}}
}

