% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius.R
\name{genius_get_all_songs_from_artist}
\alias{genius_get_all_songs_from_artist}
\title{Retrieve All Songs by an Artist from Genius API}
\usage{
genius_get_all_songs_from_artist(artist_id)
}
\arguments{
\item{artist_id}{A string representing the ID of the artist whose songs are to be retrieved. This parameter is required.}
}
\value{
A dataframe containing all songs by the specified artist. The dataframe includes various fields such as song ID, title, and other metadata.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if the \code{artist_id} is not provided.
}
\description{
This function retrieves a list of all songs by a specific artist from the Genius API, handling pagination automatically.
The function continues fetching songs until all pages are retrieved and returns the consolidated list in a dataframe format.
}
\examples{
\dontrun{
all_songs_df <- genius_get_all_songs_from_artist("12345")
print(all_songs_df)
}
}
