% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_annotation}
\alias{genius_get_annotation}
\title{Retrieve Annotation Details from Genius API}
\usage{
genius_get_annotation(annotation_id, text_format = "dom")
}
\arguments{
\item{annotation_id}{A string representing the ID of the annotation to retrieve. This parameter is required.}

\item{text_format}{A string specifying the format of the text returned. Options are "dom" (default), "html", or "plain".}
}
\value{
A list containing the annotation details in JSON format. The list includes various fields such as 'id', 'annotator_id', 'annotator_login', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200).
}
\description{
This function retrieves the details of a specific annotation from the Genius API.
}
\examples{
\dontrun{
genius_get_annotation(annotation_id = "12345")
}
}
