% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_artist}
\alias{genius_get_artist}
\title{Retrieve Artist Details from Genius API}
\usage{
genius_get_artist(artist_id, text_format = "dom")
}
\arguments{
\item{artist_id}{A string representing the ID of the artist to retrieve. This parameter is required.}

\item{text_format}{A string specifying the format of the text returned. Options are "dom" (default), "html", or "plain".}
}
\value{
A list containing the artist details in JSON format. The list includes various fields such as 'id', 'name', 'url', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if the \code{artist_id} is not provided.
}
\description{
This function retrieves detailed information about a specific artist from the Genius API.
}
\examples{
\dontrun{
genius_get_artist(artist_id = "12345")
}
}
