% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genius.R
\name{genius_get_artist_details_df}
\alias{genius_get_artist_details_df}
\title{Search for an Artist and Retrieve Their Details as a Dataframe}
\usage{
genius_get_artist_details_df(artist_name)
}
\arguments{
\item{artist_name}{A string representing the name of the artist to search for. This parameter is required.}
}
\value{
A dataframe containing details about the artist, including their ID, name, URL, header image URL, image URL,
verification status, and IQ score. The function will return an error if the artist is not found or if there are issues with the API request.
}
\description{
This function searches for an artist on Genius using their name and retrieves detailed information about the artist.
If multiple artists are found, the user is prompted to choose the correct artist. The function returns the artist's details
in a dataframe format.
}
\examples{
\dontrun{
artist_df <- genius_get_artist_details_df("Kendrick Lamar")
print(artist_df)
}
}
