% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_artist_songs}
\alias{genius_get_artist_songs}
\title{Retrieve Songs by Artist from Genius API}
\usage{
genius_get_artist_songs(artist_id, sort = "title", per_page = 20, page = 1)
}
\arguments{
\item{artist_id}{A string representing the ID of the artist whose songs are to be retrieved. This parameter is required.}

\item{sort}{A string specifying the sorting order of the results. Options are "title" (default), "popularity", etc.}

\item{per_page}{An integer specifying the number of results to return per page. Default is 20.}

\item{page}{An integer specifying the page number of results to return. Default is 1.}
}
\value{
A list containing the song details in JSON format. The list includes various fields such as 'id', 'title', 'primary_artist', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if the \code{artist_id} is not provided.
}
\description{
This function retrieves a list of songs by a specific artist from the Genius API, with options to sort and paginate the results.
}
\examples{
\dontrun{
genius_get_artist_songs(
  artist_id = "12345",
  sort = "popularity",
  per_page = 10,
  page = 2
)
}
}
