% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_referents}
\alias{genius_get_referents}
\title{Retrieve Referents from Genius API}
\usage{
genius_get_referents(
  created_by_id = NULL,
  song_id = NULL,
  web_page_id = NULL,
  text_format = "dom",
  per_page = NULL,
  page = NULL
)
}
\arguments{
\item{created_by_id}{An optional integer representing the ID of the user who created the referents. Default is \code{NULL}.}

\item{song_id}{An optional integer representing the ID of the song to filter referents by. Default is \code{NULL}. Must not be used with \code{web_page_id}.}

\item{web_page_id}{An optional integer representing the ID of the web page to filter referents by. Default is \code{NULL}. Must not be used with \code{song_id}.}

\item{text_format}{A string specifying the format of the text returned. Options are "dom" (default), "html", or "plain".}

\item{per_page}{An optional integer specifying the number of results per page. Default is \code{NULL}.}

\item{page}{An optional integer specifying the page of results to return. Default is \code{NULL}.}
}
\value{
A list containing the referents in JSON format. The list includes various fields such as 'id', 'annotator_id', 'annotator_login', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if both \code{song_id} and \code{web_page_id} are provided.
}
\description{
This function retrieves referents (annotations) from the Genius API based on various filter criteria.
}
\examples{
\dontrun{
genius_get_referents(song_id = 12345)
}
}
