% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_search_results}
\alias{genius_get_search_results}
\title{Retrieve Search Results from Genius API}
\usage{
genius_get_search_results(query)
}
\arguments{
\item{query}{A string representing the search query. This parameter is required.}
}
\value{
A list containing the search results in JSON format. The list includes various fields such as 'hits', 'title', 'primary_artist', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if the \code{query} parameter is not provided.
}
\description{
This function retrieves search results from the Genius API based on a provided search query.
}
\examples{
\dontrun{
genius_get_search_results(query = "Kendrick Lamar")
}
}
