% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_song}
\alias{genius_get_song}
\title{Retrieve Song Details from Genius API}
\usage{
genius_get_song(song_id, text_format = "dom")
}
\arguments{
\item{song_id}{A string representing the ID of the song to retrieve. This parameter is required.}

\item{text_format}{A string specifying the format of the text returned. Options are "dom" (default), "html", or "plain".}
}
\value{
A list containing the song details in JSON format. The list includes various fields such as 'id', 'title', 'primary_artist', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if the \code{song_id} is not provided.
}
\description{
This function retrieves detailed information about a specific song from the Genius API.
}
\examples{
\dontrun{
genius_get_song(song_id = "12345")
}
}
