% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusAPI.R
\name{genius_get_web_page}
\alias{genius_get_web_page}
\title{Retrieve Web Page Details from Genius API}
\usage{
genius_get_web_page(
  raw_annotatable_url = NULL,
  canonical_url = NULL,
  og_url = NULL
)
}
\arguments{
\item{raw_annotatable_url}{An optional string representing the raw annotatable URL of the web page to retrieve. Default is \code{NULL}.}

\item{canonical_url}{An optional string representing the canonical URL of the web page to retrieve. Default is \code{NULL}.}

\item{og_url}{An optional string representing the Open Graph URL of the web page to retrieve. Default is \code{NULL}.}
}
\value{
A list containing the web page details in JSON format. The list includes various fields such as 'id', 'url', 'title', etc.
The function will return an error if the request fails (i.e., if the HTTP status code is not 200) or if none of the URL parameters are provided.
}
\description{
This function retrieves details about a specific web page from the Genius API based on a provided URL.
}
\examples{
\dontrun{
genius_get_web_page(
  raw_annotatable_url = "https://genius.com/Some-annotatable-page"
)
}
}
