% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rap_coverMet.R
\name{get_rap_coverMeteorology_table}
\alias{get_rap_coverMeteorology_table}
\title{Query RAP Yearly Vegetation Cover & Meteorology Data}
\usage{
get_rap_coverMeteorology_table(
  aoi,
  year = NULL,
  mask = TRUE,
  nodata_flag = NA_real_
)
}
\arguments{
\item{aoi}{Area of Interest. A \code{SpatVector} object, or any spatial object
that can be converted with \code{terra::vect()}. The AOI coordinates will be
transformed to WGS84 longitude latitude (\code{"EPSG:4326"}). The AOI can be
specified using point, line and polygon geometries. Each unique feature
will be passed separately to the API. The result \code{feature} column contains
the row index of the input feature from \code{aoi}.}

\item{year}{integer. Optional. Numeric year or vector of years (1986 to last
full year). Default: \code{NULL} returns all available years.}

\item{mask}{logical. Exclude cropland, development, and water? Default:
\code{TRUE}.}

\item{nodata_flag}{numeric. Value to use for missing data. The API encodes
"NODATA" as \code{-99}. Default: \code{NA_real_} replaces \code{-99} with \code{NA}.}
}
\value{
A data.frame with yearly fractional cover data including the following
columns: \code{"year"} (cover estimate year), ``"AFG"\verb{(Annual Forb and Grass  cover),}"PFG"` (Perennial Forb and Grass cover), `"SHR"` (Shrub cover),
`"TRE"` (Tree cover), `"LTR"` (Litter cover), `"BGR"` (Bare Ground cover),
`"annualTemp"` (Annual average temperature in degrees Fahrenheit),
`"annualPrecip"` (Annual total precipitation in inches),
`"feature"` (feature ID, row number from `aoi`)
}
\description{
Retrieves remotely sensed production estimates from the Rangeland Analysis
Platform (RAP) using the \code{coverMeteorologyV3} API endpoint. This function
supports querying one or more spatial features (points, lines, or polygons)
provided as a \code{terra} \code{SpatVector} in WGS84 longitude latitude
(\code{"EPSG:4326"}).
}
\details{
For each feature - year combination, a separate request is made to the RAP
API, and results are returned as a combined \code{data.frame}. In the special case
of \code{year=NULL}) default all available years are returned in a single query.

For more information on the API and data products, see the RAP API
documentation: \url{https://rangelands.app/support/71-api-documentation}
}
\examples{
\dontshow{if (requireNamespace("terra") && isTRUE(as.logical(Sys.getenv("R_RAPR_EXTENDED_EXAMPLES", unset=FALSE)))) withAutoprint(\{ # examplesIf}

aoi <- terra::vect(data.frame(x = -119.72330, y = 36.92204),
                   geom = c('x', 'y'),
                   crs = "EPSG:4326")

# all years (year=NULL)
res <- get_rap_coverMeteorlogy_table(aoi)
str(res)

# specific year
res <- get_rap_coverMeteorology_table(aoi, year = 1992)
str(res)

# multiple specific years
res <- get_rap_coverMeteorology_table(aoi, year = 1993:2003)
str(res)

# 1 kilometer buffer around point
res <- get_rap_coverMeteorology_table(terra::buffer(aoi, 1000), year = 2004)
str(res)
\dontshow{\}) # examplesIf}
}
