% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mtp.R
\name{has_columns}
\alias{has_columns}
\title{Internal function to check if the data frame contains the as a minimum the right columns.}
\usage{
has_columns(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
TRUE or error
}
\description{
Internal function to check if the data frame contains the as a minimum the right columns.
}

