% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mtp.R
\name{has_format}
\alias{has_format}
\title{Internal function to check if the data frame columns have the right formats.}
\usage{
has_format(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
TRUE or an error
}
\description{
Internal function to check if the data frame columns have the right formats.
}

