% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_mtp.R
\name{median_mtp_list}
\alias{median_mtp_list}
\title{Produces the median \class{mtp} objects from a list of \code{mtp} objects}
\usage{
median_mtp_list(list_of_mtp)
}
\arguments{
\item{list_of_mtp}{A list of \code{mtp} objects}
}
\value{
Returns a list of median \code{mtp} objects
}
\description{
Produces the median \class{mtp} objects from a list of \code{mtp} objects
}
\examples{
# df_mtp is supplied with the package
combined01 <- combine_mtp_list(df_mtp$mtp)
plot_mtp_list(combine_mtp_list(df_mtp$mtp))

# using dplyr
library(dplyr)
combined02 <- df_mtp \%>\% dplyr::group_by(file) \%>\% dplyr::summarize(mtp = median_mtp_list(mtp)) #condensing
plot_mtp_list(combined$mtp)

}

