% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mtp.R
\name{mtp_extract}
\alias{mtp_extract}
\title{The function extracts mtp objects from a file sourced from BO (screening
data)}
\usage{
mtp_extract(data_frame)
}
\arguments{
\item{data_frame}{A data frame that must contain \code{Filename},
\code{Scanner}, \code{`Plate Setup`}, \code{Plate}, \code{Name},
\code{Hor.}, \code{Ver.}, as well as columns with numerical headings in
hours.}
}
\value{
a data frame with a list column \code{mtp} containing the \code{mtp} objects
}
\description{
The function extracts mtp objects from a file sourced from BO (screening
data)
}
\examples{
# to be made
}

