% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{mtp_feature}
\alias{mtp_feature}
\title{Find feature points for a MTP}
\usage{
mtp_feature(df_mtp, ...)
}
\arguments{
\item{df_mtp}{a data frame of class mtp}

\item{...}{parameters for the \code{\link{fun_xy}} function}
}
\value{
a data frame
}
\description{
Find feature points for a MTP
}
\examples{
# pH a 6 hrs
mtp_feature(df_mtp$mtp[[1]], x_out = 6, y_name = "value")

# time to pH 5.5
mtp_feature(df_mtp$mtp[[1]], x_out = 5.5, x_name = "value", y_name = "hours")

# reducing data set to whole hours
reduced_data <- mtp_feature(df_mtp$mtp[[1]], x_out = c(1:10), y_name = "value")
mtp_plot(reduced_data) + ggplot2::geom_point()

# Calculating features for the whole data frame of mtp's
# Make sure that each piece feed to the do() function is a mtp object.
# the . is a place holder.
library(dplyr)
df_mtp \%>\%
group_by(file, sheet, Plate) \%>\%
do(mtp_feature(.$mtp[[1]], x_out = c(1,6)))

}
\seealso{
\code{fun_xy} and \code{df_mtp}
}

