% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{mtp_pH_after_hours}
\alias{mtp_pH_after_hours}
\title{calculate pH after x hours}
\usage{
mtp_pH_after_hours(mtp, x_out = NULL)
}
\arguments{
\item{mtp}{A \code{mtp} object}

\item{x_out}{the hours in which to evaluate pH}
}
\value{
Returns a data frame
}
\description{
calculate pH after x hours
}
\examples{
pHhrs <- df_mtp$mtp[[1]]
mtp_pH_after_hours(pHhrs) # error
mtp_pH_after_hours(smooth_mtp(pHhrs))
mtp_pH_after_hours(smooth_mtp(pHhrs),x_out = c(1,6,12))

}

