% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{mtp_slopes}
\alias{mtp_slopes}
\title{finds the min and max segment slope within a time window}
\usage{
mtp_slopes(mtp, start = NULL, end = NULL)
}
\arguments{
\item{mtp}{A \code{mtp} object}

\item{start}{start a \code{scalar} defining the start of the search}

\item{end}{a \code{scalar} defining the end of the search}
}
\value{
Returns a data frame of features
}
\description{
finds the min and max segment slope within a time window
}
\examples{
slopes <- df_mtp$mtp[[1]]
slopes <- smooth_mtp(slopes)
mtp_slopes(slopes)
}

