% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{mtp_smooth}
\alias{mtp_smooth}
\title{smooth the well curves of a data frame of class mtp}
\usage{
mtp_smooth(mtp, ...)
}
\arguments{
\item{...}{parameter passed on to fun_smooth}

\item{df_mtp}{a data frame of class mtp and attribute type = "raw"}
}
\value{
a data_frame of class mtp and attribute type = "fit"
}
\description{
smooth the well curves of a data frame of class mtp
}
\examples{
# df_mtp is data frame of mtp objects (in the .$mtp column) that comes with the package
smooth <- mtp_smooth(df_mtp$mtp[[1]])
mtp_plot(smooth)
mtp_plot(smooth, y_name = "fit")
mtp_plot(smooth, y_name = "resid")
mtp_plot(smooth, x_name = "value", y_name = "resid")
smooth <- mtp_smooth(df_mtp$mtp[[1]], no_points = 100)
mtp_plot(smooth, y_name = "fit")
mtp_plot(smooth, y_name = "resid")
mtp_plot(smooth, x_name = "value", y_name = "resid")
smooth <- mtp_smooth(df_mtp$mtp[[1]], no_points = 100, fit_degree = 2)
mtp_plot(smooth, y_name = "fit")
mtp_plot(smooth, y_name = "resid")
mtp_plot(smooth, x_name = "value", y_name = "resid")
smooth <- mtp_smooth(df_mtp$mtp[[1]], no_points = 30, fit_degree = 1)
mtp_plot(smooth, y_name = "fit")
mtp_plot(smooth, y_name = "resid")
mtp_plot(smooth, x_name = "value", y_name = "resid")
# now we loop through all the mtp
list_of_mtp <- df_mtp$mtp
list_of_smooth_mtp <- purrr::map(list_of_mtp, mtp_smooth)

# To keep the structure of df_map we can do the following (be sure to have one mtp per row)
smooth <- df_mtp \%>\% group_by(file, sheet, ID, Plate) \%>\% do(smtp = mtp_smooth(.$mtp[[1]]))
smooth <- df_mtp \%>\% group_by(-mtp) \%>\% do(smtp = mtp_smooth(.$mtp[[1]]))

mtp_plot(smooth$smtp[[1]], y_name = "fit")
}
\seealso{
\code{fun_smooth}
}

