% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtp.R
\name{plot_mtp}
\alias{plot_mtp}
\title{Plots the curves of a \code{mtp} object}
\usage{
plot_mtp(mtp, x_name = "hours", y_name = "value", title = "my title")
}
\arguments{
\item{mtp}{A \code{mtp} object}

\item{x_name}{The name of the variable to go on the x-axis, defaults to
\code{hours}}

\item{y_name}{The name of the variable to go on the y-axis, defaults to
\code{value}}

\item{title}{The title of the plot, defaults to "my plot"}
}
\value{
Returns a \code{ggplot} object that can be manipulated or printed
}
\description{
Plots the curves of a \code{mtp} object
}
\examples{
# df_mtp is a data frame of class mtp that comes with the package
str(df_mtp)

# plot a specific mtp object
df_plot <- df_mtp$mtp[[1]]
df_plot$Name <- as.character(1:96)
plot_mtp(df_plot, title = "a MTP data frame")
}
\seealso{
\code{\link{plot_mtp_list}} for creating a list of \code{ggplot}
  objects and \code{\link{plot_to_pdf}} for plotting a list of mtp objects to
  pdf

Other plot.functions: \code{\link{plot_mtp_list}},
  \code{\link{plot_to_pdf}}
}

