% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtp.R
\name{plot_mtp_list}
\alias{plot_mtp_list}
\title{Creates a list of \code{ggplot} objects from a list of \code{mtp} objects}
\usage{
plot_mtp_list(list_of_mtp, ...)
}
\arguments{
\item{list_of_mtp}{A list of (named) \code{mtp} objects}

\item{...}{Parameters passed on to \code{\link{plot_mtp}}}
}
\value{
Returns a list of \code{ggplot} objects
}
\description{
Creates a list of \code{ggplot} objects from a list of \code{mtp} objects
}
\details{
List elements can have names, and the title of the \code{ggplot} objects defaults to
  the name of the elements in the list of \code{mtp} objects
}
\examples{
# df_mtp is a data frame of class mtp that comes with the package
names(df_mtp$mtp) <- paste("mtp", seq(1:length(mtps)))
names(df_mtp$mtp)
plots <- plot_mtp_list(df_mtp$mtp)

# to print all the plots individually using the \\code{purrr} library
purrr::walk(plots, print)

}
\seealso{
\code{\link{plot_mtp}} for plotting a \code{mtp} object and
  \code{\link{plot_to_pdf}} for plotting a list of mtp objects to pdf

Other plot.functions: \code{\link{plot_mtp}},
  \code{\link{plot_to_pdf}}
}

