% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapr-package.R
\docType{package}
\name{rapr-package}
\alias{rapr}
\alias{rapr-package}
\title{rapr: Interface to Rangeland Analysis Platform (RAP) Vegetation Biomass and Cover Products}
\description{
Provides access to Rangeland Analysis Platform (RAP) products \url{https://rangelands.app/products} for arbitrary extents via GDAL virtual file system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rapr}
  \item \url{https://humus.rocks/rapr}
  \item \url{http://humus.rocks/rapr/}
  \item Report bugs at \url{https://github.com/brownag/rapr/issues}
}

}
\keyword{internal}
