% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_169_long}
\alias{save_169_long}
\title{saves for CH 16:9 template in long format}
\usage{
save_169_long(filename = "plot.png", plot = last_plot())
}
\arguments{
\item{filename}{defaults to plot.png}

\item{plot}{- specify the ggplot object to save. Defaults to last_plot()}
}
\description{
saves for CH 16:9 template in long format
}

