% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_A4_hor}
\alias{save_A4_hor}
\title{saves for A4 horizontal template in square format}
\usage{
save_A4_hor(filename = "plot.png", plot = last_plot())
}
\arguments{
\item{filename}{defaults to plot.png}

\item{plot}{- specify the ggplot object to save. Defaults to last_plot()}
}
\description{
saves for A4 horizontal template in square format
}

