% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{smooth_all_mtp}
\alias{smooth_all_mtp}
\title{A function to smooth a list of data frames of the mtp type}
\usage{
smooth_all_mtp(list_df_mtp, ...)
}
\arguments{
\item{list_df_mtp}{A list of data frames of class mtp}

\item{...}{Smoother parameters}
}
\value{
a list of smoothed data frames of class mtp
}
\description{
A function to smooth a list of data frames of the mtp type
}
\examples{
ds <- smooth_all_mtp(dl)

}

