% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{smooth_mtp}
\alias{smooth_mtp}
\title{Smooths the individual curves of a \code{mtp} object.}
\usage{
smooth_mtp(mtp, ...)
}
\arguments{
\item{...}{parameter passed on to \code{\link{fun_smooth}}}

\item{df_mtp}{A \code{mtp} object}
}
\value{
a \code{mtp} object
}
\description{
Smooths the individual curves of a \code{mtp} object.
}
\examples{
# df_mtp is data frame of mtp objects (in the .$mtp column) that comes with the package
smooth <- smooth_mtp(df_mtp$mtp[[1]])
plot_mtp(smooth)
plot_mtp(smooth, y_name = "fit")
plot_mtp(smooth, y_name = "resid")
plot_mtp(smooth, x_name = "value", y_name = "resid")
smooth <- smooth_mtp(df_mtp$mtp[[1]], no_points = 100)
plot_mtp(smooth, y_name = "resid")
smooth <- smooth_mtp(df_mtp$mtp[[1]], no_points = 100, fit_degree = 2)
mtp_plot(smooth, y_name = "resid")
}
\seealso{
\code{\link{smooth_mtp_list}} for lists of mtp objects and \code{\link{fun_smooth}} for individual curves

Other smoothing.functions: \code{\link{fun_smooth}},
  \code{\link{smooth_mtp_list}}
}

