% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{smooth_mtp_list}
\alias{smooth_mtp_list}
\title{A function to smooth a list of data frames of the mtp type}
\usage{
smooth_mtp_list(list_of_mtp, ...)
}
\arguments{
\item{list_of_mtp}{A list of data frames of class mtp}

\item{...}{Parameters passed on to \code{\link{fun_smooth}}}
}
\value{
Returns a list of smoothed mtp objects
}
\description{
A function to smooth a list of data frames of the mtp type
}
\examples{
ds <- smooth_mtp_list(df_mtp$mtp)
ds <- smooth_mtp_list(df_mtp$mtp, no_points = 40)
plot_mtp_list(ds, y_name = "fit")
}
\seealso{
\code{\link{smooth_mtp}} for mtp objects and \code{\link{fun_smooth}} for individual curves

Other smoothing.functions: \code{\link{fun_smooth}},
  \code{\link{smooth_mtp}}
}

