% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_import.R
\name{xls_folder_read}
\alias{xls_folder_read}
\title{Reads \code{.xls} or \code{.xlsx} files containing \code{RAP} data from a given directory on disk.}
\usage{
xls_folder_read(my_path = dirname(file.choose()))
}
\arguments{
\item{my_path}{}
}
\value{
Returns a data frame of format \code{file}, \code{sheet}, \code{ID}, \code{Plate}, and \code{mtp} which contains the \code{mtp} plate objects.
}
\description{
Reads \code{.xls} or \code{.xlsx} files containing \code{RAP} data from a given directory on disk.
}
\examples{
# interactively read a folder
df_mtp <- xls_folder_read()
df_mtp
## NOT RUN
df_mtp <- xls_folder_read("./inst/extdata")
df_mtp
## END NOT RUN
}
\seealso{
\code{\link{xls_read}}
}

