% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_import.R
\name{xls_read}
\alias{xls_read}
\title{Reads a single \code{.xls} or \code{.xlsx} file containing \code{RAP} data from disk.}
\usage{
xls_read(path = file.choose(), sheets = readxl::excel_sheets(path))
}
\arguments{
\item{path}{The path of the file containing the mtp data. Defaults to file.choose()}

\item{sheets}{The name of the sheet containing the mtp data. Defaults to readxl::excel_sheets(path)}
}
\value{
Returns a data frame of format \code{file}, \code{sheet}, \code{ID}, \code{Plate}, and \code{mtp} which contains the \code{mtp} plate objects.
}
\description{
Reads a single \code{.xls} or \code{.xlsx} file containing \code{RAP} data from disk.
}
\examples{
# specifying the file path. \\code{system.file} is a function to search e.g. within a package.
df_mtp <- xls_read(path = system.file("extdata", "crystalban_salt_01.xls", package = "RAPr"))
df_mtp
df_mtp$mtp[1] # a list item containing a mtp object
class(df_mtp$mtp[1])
df_mtp$mtp[[1]] # a mtp object
class(df_mtp$mtp[[1]])

# browsing for files in the current directory
df_mtp <- read_mtp_xls()

# reading all files in a directory starting from the working directory
# replace with your own path to work
## NOT RUN
my_path <- "./inst/extdata/"
files <- list.files(path = my_path)
df_mtp <- purrr::map_df(files, function(file_name) xls_read(path = paste(my_path, file_name, sep = "")))
df_mtp
## END NOT RUN
}
\seealso{
\code{\link{xls_folder_read}}
}

