% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth.R
\name{fun_smooth}
\alias{fun_smooth}
\title{Loess smothing of a well curve}
\usage{
fun_smooth(df, x_name = "hours", y_name = "value", no_points = 40,
  fit_degree = 1)
}
\arguments{
\item{df}{takes a data frame of class mtp as input}

\item{x_name}{the name of the x variable of type character}

\item{y_name}{the name of the y variable of type character}

\item{no_points}{The number of points in the smooth window of type numeric}

\item{fit_degree}{The degree of the smooth polynomial of type numeric. must be less than or equal to 2}
}
\value{
returns a data frame of class smooth
}
\description{
Loess smothing of a well curve
}
\examples{
df <- dl[[1]]
df <- subset(df, Hor. == "A" & Ver. == "1")
fun_smooth(df)

}

