% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rap.R
\name{get_rap}
\alias{get_rap}
\title{Get Rangeland Analysis Platform (RAP) Grids}
\usage{
get_rap(
  x,
  years = 1986:2021,
  filename,
  dataset = c("vegetation-biomass", "vegetation-cover"),
  version = c("v3", "v2")
)
}
\arguments{
\item{x}{Target extent. Derived from an sf, terra or sp object or numeric vector containing xmin, ymax, xmax, ymin in WGS84 latitude/longitude decimal degrees.}

\item{years}{integer. Year(s) to query}

\item{filename}{Output filename}

\item{dataset}{Target data: either `"vegetation-biomass"` or `"vegetation-cover"`}

\item{version}{Target version: either `"v3"` or `"v2"`}
}
\description{
Get Rangeland Analysis Platform (RAP) Grids
}
