% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rap.R
\name{get_rap}
\alias{get_rap}
\title{Get Rangeland Analysis Platform (RAP) Grids}
\usage{
get_rap(
  x,
  years,
  product,
  filename = NULL,
  ...,
  source = "rap-30m",
  version = "v3",
  legacy = TRUE,
  progress = NULL
)
}
\arguments{
\item{x}{Target extent. Derived from an sf, terra, raster or sp object or
numeric vector containing xmin, ymax, xmax, ymin in WGS84
longitude/latitude decimal degrees (EPSG:4326).}

\item{years}{integer. Year(s) to query}

\item{product}{Target data: \code{"vegetation-biomass"} and/or
\code{"vegetation-cover"} (for \code{"rap-30m"}) and \code{"pft"} (plant functional type
cover), \code{"gap"} (canopy gap), \code{"arte"} (Artemisia spp. cover), \code{"iag"}
(invasive annual grass cover), \code{"pj"} (pinyon juniper cover))}

\item{filename}{Output filename (optional; default stores in temporary files,
see \code{terra::sources()})}

\item{...}{Additional arguments passed to internal RAP query function and
\verb{[terra::writeRaster()]}}

\item{source}{Grid sources. Options include \code{"rap-30m"} (default; Landsat)
and \code{"rap-10m"} (Sentinel 2).}

\item{version}{Target version: \code{"v3"} and/or \code{"v2"} (for \verb{"rap-30m}). Ignored
for \code{"rap-10m"}.}

\item{legacy}{\emph{logical}. Use legacy (gdal_translate) method? Default: \code{TRUE}
(applies only to \code{source="rap-30m"}).}

\item{progress}{logical. Show progress bar? Default: missing (\code{NULL}) will
use progress bar when three or more layers are requested.}
}
\value{
a SpatRaster containing the requested product layers by year.
Native cell resolution of \code{"rap-30m"} is ~30m x 30m in WGS84 geographic coordinate system.
Native cell resolution of \code{"rap-10m"} is 10m x 10m in WGS84 Universal Transverse Mercator (UTM) zone.
}
\description{
Get Rangeland Analysis Platform (RAP) Grids
}
\details{
You can query annual biomass and cover (versions 2 and 3) from 1986
to present
\itemize{
\item \code{product = "vegetation-biomass"} returns two layers per year:
\itemize{
\item \code{"annual forb and grass"}, \code{"perennial forb and grass"} (\strong{lbs / acre})
}
\item \code{product = "vegetation-cover"} returns six layers per year:
\itemize{
\item \code{"annual forb and grass"}, \code{"bare ground"}, \code{"litter"}, \code{"perennial forb and grass"}, \code{"shrub"}, \code{"tree"} (\strong{\% cover})
}
}

When a \code{filename} argument is not specified, unique temporary files will be
generated. The resulting SpatRaster object will retain reference to these
files, and you can remove them manually with
\verb{unlink(terra::sources(<SpatRaster))}.

When a \code{filename} \emph{is} specified, temporary files will be removed after the
result (often a multi- year/layer/product) SpatRaster is written to new
file.

In lieu of a spatial object from \{terra\}, \{raster\}, \{sf\} or \{sp\}
packages you may specify a bounding box using a numeric vector containing
\code{xmin}, \code{ymax}, \code{xmax}, \code{ymin} in WGS84 longitude/latitude decimal degrees
(corresponding to order used in \code{gdal_translate} \code{-projwin} option). e.g.
\code{get_rap(x = c(-120, 37, -119.99, 36.99), ...)}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{(1: xmin, 2: ymax)--------------------------|
        |                                   |
        |         TARGET EXTENT             |
        |  x = c(xmin, ymax, xmax, ymin)    |
        |                                   |
        |---------------------------(3: xmax, 4: ymin)
}\if{html}{\out{</div>}}
}
