% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rap_table.R
\name{get_rap_table}
\alias{get_rap_table}
\title{Query RAP Tabular Data}
\usage{
get_rap_table(
  aoi,
  years = NULL,
  product,
  version = "V3",
  mask = TRUE,
  nodata_flag = NA_real_
)
}
\arguments{
\item{aoi}{Area of Interest. A \code{SpatVector} object, or any spatial object
that can be converted with \code{terra::vect()}. The AOI coordinates will be
transformed to WGS84 longitude latitude (\code{"EPSG:4326"}). The AOI can be
specified using point, line and polygon geometries. Each unique feature
will be passed separately to the API. The result \code{feature} column contains
the row index of the input feature from \code{aoi}.}

\item{years}{integer. Optional. Numeric year or vector of years (1986 to last
full year). Default: \code{NULL} returns all available years.}

\item{product}{Target data: \code{"cover"}, \code{"coverMeteorology"},
\code{"production"}, and/or \code{"production16day"}.}

\item{version}{Target version: \code{"v3"}. Currently ignored.}

\item{mask}{logical. Exclude cropland, development, and water? Default:
\code{TRUE}.}

\item{nodata_flag}{numeric. Value to use for missing data. The API encodes
"NODATA" as \code{-99}. Default: \code{NA_real_} replaces \code{-99} with \code{NA}.}
}
\value{
A \emph{data.frame} with requested time-series data by year or 16-day
production period. Columns included for each product option are described in
Details above.
}
\description{
Retrieves remotely sensed production & cover estimates from the Rangeland
Analysis Platform (RAP) using the tabular data API endpoints. This function
supports querying one or more spatial features (points, lines, or polygons)
provided as a \code{terra} \code{SpatVector} object, or any spatial object that can be
converted with \code{terra::vect()}. See Details for the products available.
}
\details{
For each feature - year combination, a separate request is made to the RAP
API, and results are returned as a combined \code{data.frame}. In the special case
of \code{years=NULL}) default all available years are returned in a single query.

For more information on the API and data products, see the RAP API
documentation: \url{https://rangelands.app/support/71-api-documentation}
\subsection{Products Information}{

You can query several Landsat derived biomass, cover, and meteorological products from 1986 to present:
\itemize{
\item \code{product = "cover"} returns a data.frame with yearly fractional cover data including the following columns:
\itemize{
\item \code{"year"} (cover estimate year), \code{"AFG"} (Annual Forb and Grass cover),
\code{"PFG"} (Perennial Forb and Grass cover), \code{"SHR"} (Shrub cover),
\code{"TRE"} (Tree cover), \code{"LTR"} (Litter cover), \code{"BGR"} (Bare Ground cover),
\code{"feature"} (feature ID, row number from \code{aoi}) (\strong{\% cover})
}
\item \code{product = "coverMeteorology"} returns a data.frame with yearly fractional cover & meteorological data including the following columns:
}

\code{"year"} (cover estimate year), \code{"AFG"} (Annual Forb and Grass cover),
\code{"PFG"} (Perennial Forb and Grass cover), \code{"SHR"} (Shrub cover),
\code{"TRE"} (Tree cover), \code{"LTR"} (Litter cover), \code{"BGR"} (Bare Ground cover),
\code{"annualTemp"} (Annual average temperature in degrees Fahrenheit),
\code{"annualPrecip"} (Annual total precipitation in inches),
\code{"feature"} (feature ID, row number from \code{aoi}) (\strong{\% cover})
\itemize{
\item \code{product = "production"} returns a data.frame with yearly production data including the following columns:
}

\code{"year"} (production estimate year), \code{"AFG"} (Annual Forb and Grass production),
\code{"PFG"} (Perennial Forb and Grass production), \code{"HER"} (Herbaceous production),
\code{"feature"} (feature ID, row number from \code{aoi}) (\strong{lbs / acre})
\itemize{
\item \code{product = "production16day"} returns a data.frame with 16-day production data including the following columns:
}

\code{"date"} (production estimate date), \code{"year"} (production estimate year),
\code{"doy"} (production estimate Julian day of year), \code{"AFG"} (Annual Forb and
Grass production), \code{"PFG"} (Perennial Forb and Grass production), \code{"HER"}
(Herbaceous production), \code{"feature"} (feature ID, row number from \code{aoi})
(\strong{lbs / acre})
}
}
\examples{
\dontshow{if (requireNamespace("terra") && isTRUE(as.logical(Sys.getenv("R_RAPR_EXTENDED_EXAMPLES", unset=FALSE)))) withAutoprint(\{ # examplesIf}

aoi <- terra::vect(data.frame(x = -119.72330, y = 36.92204),
                   geom = c('x', 'y'),
                   crs = "EPSG:4326")

# all years (years=NULL)
res <- get_rap_table(aoi)
str(res)

# specific year
res <- get_rap_table(aoi, years = 1992)
str(res)

# multiple specific years
res <- get_rap_table(aoi, years = 1993:2003)
str(res)

# 1 kilometer buffer around point
res <- get_rap_table(terra::buffer(aoi, 1000), years = 2004)
str(res)

# all years (years=NULL) 16 day production data using dedicated function 
res <- get_rap_production16day_table(aoi)
str(res)
\dontshow{\}) # examplesIf}
}
