% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtp.R
\name{plot_to_pdf}
\alias{plot_to_pdf}
\title{Takes a list of \code{mtp} objects and plots them to pdf}
\usage{
plot_to_pdf(list_of_mtp, file_name = "my_plots.pdf", ...)
}
\arguments{
\item{list_of_mtp}{A list of \code{mtp} objects}
}
\value{
nothing but saves a pdf files of \code{ggplot}'s
}
\description{
Takes a list of \code{mtp} objects and plots them to pdf
}
\examples{
library(dplyr)
df_plots <- df_mtp \%>\%
transmute(title = paste("file", .$file[1], "sheet", .$sheet[1], "plate", .$Plate[1]), mtp = mtp)

dl <- df_plots$mtp
names(dl) <- df_plots$title

plot_to_pdf(dl)

}
\seealso{
\code{\link{plot_mtp}} for plotting a \code{mtp} object and
  \code{\link{plot_mtp_list}} for making a list of \code{ggplot} objects

Other plot.functions: \code{\link{plot_mtp_list}},
  \code{\link{plot_mtp}}
}

