% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapr-package.R
\docType{package}
\name{rapr-package}
\alias{rapr}
\alias{rapr-package}
\title{rapr: Interface to 'Rangeland Analysis Platform' (RAP) Products}
\description{
Provides access to 'Rangeland Analysis Platform' (RAP) products \url{https://rangelands.app/products} for arbitrary extents via 'GDAL' virtual file system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rapr}
  \item \url{https://humus.rocks/rapr}
  \item Report bugs at \url{https://github.com/brownag/rapr/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

Other contributors:
\itemize{
  \item Georgia Harrison (\href{https://orcid.org/0000-0003-0544-4471}{ORCID}) [contributor]
}

}
\keyword{internal}
