% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapr-package.R
\docType{package}
\name{rapr-package}
\alias{rapr}
\alias{rapr-package}
\title{rapr: Interface to Rangeland Analysis Platform (RAP) Vegetation Biomass and Cover Products}
\description{
Provides access to Rangeland Analysis Platform (RAP) products <https://rangelands.app/products> for arbitrary extents via GDAL virtual file system.
}
\keyword{internal}
